added jwtoken as part of header in middleware/auth.js  file
